backend/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
librosa
praat-parselmouth
python-multipart
pandas
ffmpeg-python
pydantic